Separate incident data type into its own file
