+++ site/gatsby-site/server/fields/incidents.ts
@@ -9,6 +9,7 @@ import { linkReportsToIncidents } from "./common";
 import { UserType } from "../types/user";
 import { ReportType } from "../types/report";
 import { EntityType } from "../types/entity";
+import { IncidentType } from "../types/incidents";
 
 
 export const incidentEmbedding = (reports: Record<string, any>[]) => {
@@ -27,80 +28,6 @@ export const incidentEmbedding = (reports: Record<string, any>[]) => {
             from_reports: reports.map((report) => report.report_number),
         };
 };
-
-
-const EmbeddingType = new GraphQLObjectType({
-    name: 'IncidentEmbedding',
-    fields: {
-        from_reports: { type: new GraphQLList(GraphQLInt) },
-        vector: { type: new GraphQLList(GraphQLFloat) }
-    }
-});
-
-const TsneType = new GraphQLObjectType({
-    name: 'IncidentTsne',
-    fields: {
-        x: { type: GraphQLFloat },
-        y: { type: GraphQLFloat }
-    }
-});
-
-const IncidentType = new GraphQLObjectType({
-    name: 'Incident',
-    fields: {
-        _id: { type: ObjectIdScalar },
-        date: { type: new GraphQLNonNull(GraphQLString) },
-        description: { type: GraphQLString },
-        editor_notes: { type: GraphQLString },
-        epoch_date_modified: { type: GraphQLInt },
-        incident_id: { type: new GraphQLNonNull(GraphQLInt) },
-        title: { type: new GraphQLNonNull(GraphQLString) },
-        AllegedDeployerOfAISystem: {
-            type: new GraphQLList(EntityType),
-            resolve: getListRelationshipResolver('AllegedDeployerOfAISystem', 'entity_id', EntityType, 'aiidprod', 'entities', 'Alleged deployer of AI system'),
-            extensions: {
-                relationship: getListRelationshipExtension('AllegedDeployerOfAISystem', 'entity_id', GraphQLString, 'aiidprod', 'entities')
-            },
-        },
-        AllegedDeveloperOfAISystem: {
-            type: new GraphQLList(EntityType),
-            resolve: getListRelationshipResolver('AllegedDeveloperOfAISystem', 'entity_id', EntityType, 'aiidprod', 'entities', 'Alleged developer of AI system'),
-            extensions: {
-                relationship: getListRelationshipExtension('AllegedDeveloperOfAISystem', 'entity_id', GraphQLString, 'aiidprod', 'entities')
-            },
-        },
-        AllegedHarmedOrNearlyHarmedParties: {
-            type: new GraphQLList(EntityType),
-            resolve: getListRelationshipResolver('AllegedHarmedOrNearlyHarmedParties', 'entity_id', EntityType, 'aiidprod', 'entities', 'Alleged harmed or nearly harmed parties'),
-            extensions: {
-                relationship: getListRelationshipExtension('AllegedHarmedOrNearlyHarmedParties', 'entity_id', GraphQLString, 'aiidprod', 'entities')
-            },
-        },
-        editor_dissimilar_incidents: { type: new GraphQLList(GraphQLInt) },
-        editor_similar_incidents: { type: new GraphQLList(GraphQLInt) },
-        editors: getListRelationshipConfig(UserType, GraphQLString, 'editors', 'userId', 'users', 'customData'),
-        embedding: { type: EmbeddingType },
-        flagged_dissimilar_incidents: { type: new GraphQLList(GraphQLInt) },
-        nlp_similar_incidents: { type: new GraphQLList(NlpSimilarIncidentType) },
-        reports: getListRelationshipConfig(ReportType, GraphQLInt, 'reports', 'report_number', 'reports', 'aiidprod'),
-        tsne: { type: TsneType }
-    },
-});
-
-// dependencies property gets ignored by newest graphql package so we have to add it manually after the type is created
-// ideally should be fixed in the graphql-to-mongodb package
-
-//@ts-ignore 
-IncidentType.getFields().reports.dependencies = ['reports'];
-//@ts-ignore 
-IncidentType.getFields().AllegedDeployerOfAISystem.dependencies = ['Alleged deployer of AI system'];
-//@ts-ignore 
-IncidentType.getFields().AllegedDeveloperOfAISystem.dependencies = ['Alleged developer of AI system'];
-//@ts-ignore 
-IncidentType.getFields().AllegedHarmedOrNearlyHarmedParties.dependencies = ['Alleged harmed or nearly harmed parties'];
-//@ts-ignore 
-IncidentType.getFields().editors.dependencies = ['editors'];
-
 
 export const queryFields: GraphQLFieldConfigMap<any, Context> = {
